Refactor: merge recoil atoms
recoil atom들을 store로 이동 시켰습니다.
