+++ frontend/src/organism/ChannelHeader/ChannelHeader.js
@@ -1,8 +1,7 @@
 import React, { useState, useEffect } from 'react'
-import { useHistory } from 'react-router'
-import request from '../../util/request'
 import styled from 'styled-components'
 import { toast } from 'react-toastify'
+import { useHistory } from 'react-router'
 import Icon from '../../atom/Icon'
 import { ADDUSER, INFOCIRCLE } from '../../constant/icon'
 import ChannelCard from '../../atom/ChannelCard'
@@ -10,7 +9,7 @@ import ChannelStarBtn from '../../atom/ChannelStarBtn'
 import ChannelPinBtn from '../../atom/ChannelPinBtn'
 import ChannelTopicBtn from '../../atom/ChannelTopicBtn'
 import ChannelMemberThumbnail from '../../atom/ChannelMemberThumbnail'
-import modalAtom from '../../recoil/modalAtom'
+import { modalAtom } from '../../store'
 import { useRecoilState } from 'recoil'
 import InviteUserToChannelModal from '../InviteUserToChannelModal'
 import useCurrentChannelInfo from '../../hooks/useCurrentChannelInfo'
