--- frontend/src/recoil/channelInfoAtom.js
@@ -1,8 +1,0 @@
-import { atom } from 'recoil'
-
-const channelInfoAtom = atom({
-  key: 'ChannelInfo',
-  default: {},
-})
-
-export default channelInfoAtom
