--- frontend/src/recoil/modalAtom.js
@@ -1,8 +1,0 @@
-import { atom } from 'recoil'
-
-const modalAtom = atom({
-  key: 'Modal',
-  default: null,
-})
-
-export default modalAtom
