--- frontend/src/recoil/workspaceUserInfoAtom.js
@@ -1,8 +1,0 @@
-import { atom } from 'recoil'
-
-const workspaceUserInfoAtom = atom({
-  key: 'WorkspaceUserInfo',
-  default: {},
-})
-
-export default workspaceUserInfoAtom
